saluki_core::components::transforms

Trait SynchronousTransform

pub trait SynchronousTransform {
    // Required method
    fn transform_buffer(&mut self, buffer: &mut EventsBuffer);
}

A synchronous transform.

Synchronous transforms are conceptually similar to transforms, designed to manipulate events. However, in some cases, it can be beneficial to perform multiple transformation operations as a single step, without the overhead of individual transform components. Synchronous transforms allow discrete transformation logic to be combined together within a single transform component for processing efficiency.

Required Methods

fn transform_buffer(&mut self, buffer: &mut EventsBuffer)

Transforms the events in the event buffer.