saluki_core::components::transforms

Trait SynchronousTransformBuilder

pub trait SynchronousTransformBuilder: MemoryBounds {
    // Required method
    fn build<'life0, 'async_trait>(
        &'life0 self,
        context: ComponentContext,
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn SynchronousTransform + Send>, GenericError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

A synchronous transform builder.

Synchronous transforms are a special case of transforms that are executed synchronously, meaning that they do not run as their own task. This is used for certain transforms, typically those where it is more efficient to run multiple transformation steps in a single task.

Required Methods

fn build<'life0, 'async_trait>( &'life0 self, context: ComponentContext, ) -> Pin<Box<dyn Future<Output = Result<Box<dyn SynchronousTransform + Send>, GenericError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Builds an instance of the synchronous transform.

The provided context is relative to the parent component that is building this synchronous transform.

Errors

If the synchronous transform cannot be built for any reason, an error is returned.