saluki_core::data_model::event

Module eventd

Events.

Structs

EventD

EventD is an object that can be posted to the DataDog event stream.

Enums

AlertType

Alert type.

Priority

Event priority.

Constants

ALERT_TYPE_ERROR

Value used to specify an error alert.

ALERT_TYPE_SUCCESS

Value used to specify a success alert.

ALERT_TYPE_WARNING

Value used to specify a warning alert.

PRIORITY_LOW

Value supplied used to specify a low priority event