saluki_core::data_model::event::eventd

Struct EventD

pub struct EventD { /* private fields */ }

EventD is an object that can be posted to the DataDog event stream.

Implementations

impl EventD

pub fn title(&self) -> &str

Returns the title of the event.

pub fn text(&self) -> &str

Returns the text of the event.

pub fn hostname(&self) -> Option<&str>

Returns the host where the event originated from.

pub fn aggregation_key(&self) -> Option<&str>

Returns the aggregation key of the event.

pub fn priority(&self) -> Option<Priority>

Returns the priority of the event.

pub fn source_type_name(&self) -> Option<&str>

Returns the source type name of the event.

pub fn alert_type(&self) -> Option<AlertType>

Returns the alert type of the event.

pub fn timestamp(&self) -> Option<u64>

Returns the timestamp of the event.

This is a Unix timestamp, or the number of seconds since the Unix epoch.

pub fn tags(&self) -> &SharedTagSet

Returns the tags associated with the event.

pub fn origin_tags(&self) -> &SharedTagSet

Returns the origin tags associated with the event.

pub fn with_timestamp(self, timestamp: impl Into<Option<u64>>) -> Self

Set the timestamp.

Represented as a Unix timestamp, or the number of seconds since the Unix epoch.

This variant is specifically for use in builder-style APIs.

pub fn set_timestamp(&mut self, timestamp: impl Into<Option<u64>>)

Set the timestamp.

Represented as a Unix timestamp, or the number of seconds since the Unix epoch.

pub fn with_hostname(self, hostname: impl Into<Option<MetaString>>) -> Self

Set the hostname where the event originated from.

This variant is specifically for use in builder-style APIs.

pub fn set_hostname(&mut self, hostname: impl Into<Option<MetaString>>)

Set the hostname where the event originated from.

pub fn with_aggregation_key( self, aggregation_key: impl Into<Option<MetaString>>, ) -> Self

Set the aggregation key of the event.

Aggregation key is use to group events together in the event stream.

This variant is specifically for use in builder-style APIs.

pub fn set_aggregation_key( &mut self, aggregation_key: impl Into<Option<MetaString>>, )

Set the hostname where the event originated from.

Aggregation key is use to group events together in the event stream.

pub fn with_priority(self, priority: impl Into<Option<Priority>>) -> Self

Set the priority of the event.

This variant is specifically for use in builder-style APIs.

pub fn set_priority(&mut self, priority: impl Into<Option<Priority>>)

Set the priority of the event.

pub fn with_source_type_name( self, source_type_name: impl Into<Option<MetaString>>, ) -> Self

Set the source type name of the event.

This variant is specifically for use in builder-style APIs.

pub fn set_source_type_name( &mut self, source_type_name: impl Into<Option<MetaString>>, )

Set the source type name of the event.

pub fn with_alert_type(self, alert_type: impl Into<Option<AlertType>>) -> Self

Set the alert type of the event.

This variant is specifically for use in builder-style APIs.

pub fn set_alert_type(&mut self, alert_type: impl Into<Option<AlertType>>)

Set the alert type name of the event.

pub fn with_tags(self, tags: impl Into<SharedTagSet>) -> Self

Set the tags of the event.

This variant is specifically for use in builder-style APIs.

pub fn set_tags(&mut self, tags: impl Into<Option<SharedTagSet>>)

Set the tags of the event.

pub fn with_origin_tags(self, origin_tags: SharedTagSet) -> Self

Set the origin tags of the event.

This variant is specifically for use in builder-style APIs.

pub fn new(title: impl Into<MetaString>, text: impl Into<MetaString>) -> Self

Creates an EventD from the given title and text.

Defaults to an informational alert with normal priority.