saluki_core::data_model::event::metric

Struct Metric

pub struct Metric { /* private fields */ }

A metric.

Metrics represent the measurement of a particular quantity at a particular point in time. Several different metric types exist that provide different views into the underlying quantity: counters for representing the quantities that are aggregated/totaled over time, gauges for tracking the latest value of a quantity, and histograms for tracking the distribution of a quantity.

Structure

A metric is composed of three parts: the context, the value, and the metadata.

The context represents the โfullโ name of the metric, which includes not only the name (e.g. http_requests_total), but the tags as well. Effectively, a context is meant to be a unique name for a metric.

The value is precisely what it sounds like: the value of the metric. The value holds both the metric type and the measurement (or measurements) tied to that metric type. This ensures that the measurement(s) are always represented correctly for the given metric type.

The metadata contains ancillary data related to the metric, such as the timestamp, sample rate, and origination information like hostname and sender.

Implementations

impl Metric

pub fn counter<C, V>(context: C, values: V) -> Self
where C: Into<Context>, V: Into<ScalarPoints>,

Creates a counter metric from the given context and value(s).

Default metadata will be used.

pub fn gauge<C, V>(context: C, values: V) -> Self
where C: Into<Context>, V: Into<ScalarPoints>,

Creates a gauge metric from the given context and value(s).

Default metadata will be used.

pub fn rate<C, V>(context: C, values: V, interval: Duration) -> Self
where C: Into<Context>, V: Into<ScalarPoints>,

Creates a rate metric from the given context and value(s).

Default metadata will be used.

pub fn set<C, V>(context: C, values: V) -> Self
where C: Into<Context>, V: Into<SetPoints>,

Creates a set metric from the given context and value(s).

Default metadata will be used.

pub fn histogram<C, V>(context: C, values: V) -> Self
where C: Into<Context>, V: Into<HistogramPoints>,

Creates a histogram metric from the given context and value(s).

Default metadata will be used.

pub fn distribution<C, V>(context: C, values: V) -> Self
where C: Into<Context>, V: Into<SketchPoints>,

Creates a distribution metric from the given context and value(s).

Default metadata will be used.

pub fn context(&self) -> &Context

Gets a reference to the context.

pub fn context_mut(&mut self) -> &mut Context

Gets a mutable reference to the context.

pub fn values(&self) -> &MetricValues

Gets a reference to the values.

pub fn values_mut(&mut self) -> &mut MetricValues

Gets a mutable reference to the values.

pub fn metadata(&self) -> &MetricMetadata

Gets a reference to the metadata.

pub fn metadata_mut(&mut self) -> &mut MetricMetadata

Gets a mutable reference to the metadata.

pub fn into_parts(self) -> (Context, MetricValues, MetricMetadata)

Consumes the metric and returns the individual parts.

pub fn from_parts( context: Context, values: MetricValues, metadata: MetricMetadata, ) -> Self

Creates a Metric from the given parts.