saluki_core

Macro pooled_newtype

macro_rules! pooled_newtype {
    (outer => $name:ident, inner => $inner_ty:ty $(,)?) => { ... };
}

Creates a struct that can be stored in an object pool, based on an existing struct definition.

In order to store a value in an object pool, the item must implement the Poolable trait. This trait, and overall design of ObjectPool, dictates that a pooled data type actually holds an inner value, which is the value that is actually pooled, while the outer struct is simply a wrapper around the data that ensures it is returned to the object pool when no longer in use.

In many cases, the “inner value” might be either an existing type that cannot be modified, or there might be a need to define that struct further, such as defining struct methods or trait implementations which would be cumbersome/confusing to do on the auto-generated inner type from pooled!. In these cases, pooled_newtype! provides the simplest possible wrapper over an existing struct definition to create a Poolable version.

Implementors are required to define their data struct, including an implementation of Clearable, and then use pooled_newtype! to wrap it.

Usage

use saluki_core::{pooled_newtype, pooling::Clearable};

pub struct PreallocatedByteBuffer {
    data: Vec<u8>,
}

impl PreallocatedByteBuffer {
    pub fn new() -> Self {
        Self {
            data: Vec::with_capacity(1024)
        }
    }
}

impl Clearable for PreallocatedByteBuffer {
   fn clear(&mut self) {
      self.data.clear();
   }
}

pooled_newtype! {
   outer => ByteBuffer,
   inner => PreallocatedByteBuffer,
}

// This creates a new struct called `ByteBuffer` which simply wraps over `PreallocatedByteBuffer`. We can
// define some helper methods on `ByteBuffer` to make it easier to work with:
impl ByteBuffer {
    pub fn len(&self) -> usize {
        self.data().data.len()
    }

    pub fn write_buf(&mut self, buf: &[u8]) {
        self.data_mut().data.extend_from_slice(buf);
    }
}

fn use_byte_buffer(mut buf: ByteBuffer) {
    assert_eq!(buf.len(), 0);

    buf.write_buf(b"Hello, world!");
    assert_eq!(buf.len(), 13);
}