saluki_core::observability

Module metrics

Internal metrics support.

Structs

FilterHandle

Handle to the metrics filter.

MetricsStream

Internal metrics stream

Functions

initialize_metrics

Initializes the metrics subsystem with the given metrics prefix.

Type Aliases

SharedEvents

A collection of events that can be cheaply cloned and shared between multiple consumers.