saluki_core::observability::metrics

Function get_shared_metrics_state

pub async fn get_shared_metrics_state() -> Reflector<AggregatedMetricsProcessor>

Gets the shared metrics state, which provides unified access to internal metrics in a simplified interface.

This is lazily initialized and will only be created when it’s first accessed.