saluki_core::pooling

Module helpers

Helpers for creating and working with poolable objects.

Macros

pooled

Creates a struct that can be stored in an object pool, based on an inline struct definition.

pooled_newtype

Creates a struct that can be stored in an object pool, based on an existing struct definition.

Functions

get_pooled_object_via_builder

Creates an poolable object (of type T) when T::Data implements Default.

get_pooled_object_via_default

Creates an poolable object (of type T) when T::Data implements Default.