saluki_core

Module topology

Topology building.

Modules

interconnect

Component interconnects.

shutdown

Helpers for signaling the controlled shutdown of tasks.

Structs

BuiltTopology

A built topology.

ComponentId

A component identifier.

ComponentOutputId

A component output identifier.

OutputDefinition

An output definition.

RunningTopology

A running topology.

TopologyBlueprint

A topology blueprint represents a directed graph of components.

TopologyContext

Topology context.

TypedComponentId

A component identifier that specifies the component type.

TypedComponentOutputId

Unique identifier for a specified output of a component, including the data type of the output.

Enums

BlueprintError

A topology blueprint error.

OutputName

An output name.

Type Aliases

EventsBuffer

Default type for dispatching/consuming items from event-based components.

EventsConsumer

Default consumer for event-based components.

EventsDispatcher

Default dispatcher for event-based components.

PayloadsBuffer

Default type for dispatching/consuming items from payload-based components.

PayloadsConsumer

Default consumer for payload-based components.

PayloadsDispatcher

Default dispatcher for payload-based components.