saluki_core::topology::interconnect

Struct BufferedDispatcher

pub struct BufferedDispatcher<'a, T> { /* private fields */ }

A buffered dispatcher.

BufferedDispatcher provides an efficient and ergonomic interface to Dispatcher that allows for writing events one-by-one into batches, which are then dispatched to the configured output as needed. This allows callers to focus on the logic around what items to send, without needing to worry about the details of event buffer sizing or flushing.

Implementations

impl<T> BufferedDispatcher<'_, T>
where T: DispatchBuffer,

pub async fn push(&mut self, item: T::Item) -> Result<(), GenericError>

Pushes an item into the buffered dispatcher.

Errors

If there is an error flushing items to the output, or if there is an error acquiring a new buffer, an error is returned.

pub async fn send_all<I>(self, items: I) -> Result<usize, GenericError>
where I: IntoIterator<Item = T::Item>,

Consumes this buffered dispatcher and sends/flushes all input items to the underlying output.

If flushing is successful, Ok(flushed) is returned, where flushed is the total number of items that have been flushed through this buffered dispatcher.

Errors

If there is an error sending items to the output, an error is returned.

pub async fn flush(self) -> Result<usize, GenericError>

Consumes this buffered dispatcher, flushing any buffered items to the underlying output.

If flushing is successful, Ok(flushed) is returned, where flushed is the total number of items that have been flushed through this buffered dispatcher.

Errors

If there is an error sending items to the output, an error is returned.