saluki_core::observability::metrics

Function initialize_metrics

pub async fn initialize_metrics(
    metrics_prefix: String,
) -> Result<FilterHandle, GenericError>

Initializes the metrics subsystem with the given metrics prefix.

Errors

If a global recorder was already installed, an error will be returned.